+++ forge/src/main/java/com/leon/saintsdragons/forge/client/accessor/CameraAccessor.java
@@ -1,7 +1,7 @@
 package com.leon.saintsdragons.forge.client.accessor;
 
 public interface CameraAccessor {
-    void saintsdragons$invokeMove(double distance, double yaw, double pitch);
-    double saintsdragons$invokeGetMaxZoom(double distance);
+    void saintsdragons$invokeMove(float distance, float yaw, float pitch);
+    float saintsdragons$invokeGetMaxZoom(float distance);
     void saintsdragons$invokeSetPosition(double x, double y, double z);
 }
