+++ forge/src/main/java/com/leon/saintsdragons/forge/mixin/client/CameraMixin.java
@@ -8,10 +8,10 @@ import org.spongepowered.asm.mixin.gen.Invoker;
 @Mixin(Camera.class)
 public interface CameraMixin extends CameraAccessor {
     @Invoker("move")
-    void saintsdragons$invokeMove(double distance, double yaw, double pitch);
+    void saintsdragons$invokeMove(float distance, float yaw, float pitch);
 
     @Invoker("getMaxZoom")
-    double saintsdragons$invokeGetMaxZoom(double distance);
+    float saintsdragons$invokeGetMaxZoom(float distance);
 
     @Invoker("setPosition")
     void saintsdragons$invokeSetPosition(double x, double y, double z);
